fixing admin page

+++ client/src/admin/products/AdminProduct.jsx
@@ -66,7 +66,10 @@ const AdminProduct = ({ product }) => {
         body: JSON.stringify({ quantity: updatedQuantity }),
       };
 
-      fetch(`${import.meta.env.VITE_URL}/api/product/${productId}`, requestOptions)
+      fetch(
+        `${import.meta.env.VITE_URL}/api/product/${productId}`,
+        requestOptions
+      )
         .then((response) => response.json())
         .then((result) => {
           return result.status === "success"
@@ -83,45 +86,47 @@ const AdminProduct = ({ product }) => {
     }
   };
 
-  return (
-    <article className="flex flex-col shadow">
-      <div className="flex flex-row justify-between p-2 m-1">
-        <h1 className="text-left font-bold">{product?.title}</h1>
-        <button
-          className="px-2 h-6 rounded-full bg-[#C8B8B4] text-white ease-in-out duration-300 hover:bg-white hover:text-black shadow"
-          onClick={() => deleteProduct(product?.id)}
-        >
-          X
-        </button>
-      </div>
+  if (product) {
+    return (
+      <article className="flex flex-col shadow">
+        <div className="flex flex-row justify-between p-2 m-1">
+          <h1 className="text-left font-bold">{product?.title}</h1>
+          <button
+            className="px-2 h-6 rounded-full bg-[#C8B8B4] text-white ease-in-out duration-300 hover:bg-white hover:text-black shadow"
+            onClick={() => deleteProduct(product?.id)}
+          >
+            X
+          </button>
+        </div>
 
-      <img src={product?.imageUrl} alt={"picture of product"} />
-      <p className="p-2 text-left">{product?.description}</p>
-      <p className="text-left p-2">{product?.price}</p>
-      <div className="flex flex-row p-2 items-center">
-        <p className="whitespace-nowrap font-semibold">
-          Quantity: {product?.quantity}
-        </p>
-        {update ? (
-          <>
-            <input
-              className="border rounded shadow-inner px-1 py-1 w-1/4"
-              type={"number"}
-              placeholder="e.g. 1"
-              value={updatedQuantity}
-              onChange={(e) => setUpdatedQuantity(e.target.value)}
-            />{" "}
-            <AiOutlineCheck onClick={() => updateQuantity(product?.id)} />
-          </>
-        ) : (
-          <AiOutlineEdit
-            onClick={() => setUpdate(!update)}
-            className="text-lg cursor-pointer ml-2"
-          />
-        )}
-      </div>
-    </article>
-  );
+        <img src={product?.imageUrl} alt={"picture of product"} />
+        <p className="p-2 text-left">{product?.description}</p>
+        <p className="text-left p-2">{product?.price}</p>
+        <div className="flex flex-row p-2 items-center">
+          <p className="whitespace-nowrap font-semibold">
+            Quantity: {product?.quantity}
+          </p>
+          {update ? (
+            <>
+              <input
+                className="border rounded shadow-inner px-1 py-1 w-1/4"
+                type={"number"}
+                placeholder="e.g. 1"
+                value={updatedQuantity}
+                onChange={(e) => setUpdatedQuantity(e.target.value)}
+              />{" "}
+              <AiOutlineCheck onClick={() => updateQuantity(product?.id)} />
+            </>
+          ) : (
+            <AiOutlineEdit
+              onClick={() => setUpdate(!update)}
+              className="text-lg cursor-pointer ml-2"
+            />
+          )}
+        </div>
+      </article>
+    );
+  }
 };
 
 export default AdminProduct;
